允许startBrowser设置baseUrl(通过headers里的Origin)

+++ app/src/main/java/io/legado/app/ui/book/info/BookInfoViewModel.kt
@@ -461,9 +461,11 @@ class BookInfoViewModel(application: Application) : BaseViewModel(application) {
         execute {
             bookData.value?.let {
                 it.delete()
+                if(it.coverUrl.isNullOrBlank()){
                 val future = Glide.with(context).downloadOnly().apply ( RequestOptions().onlyRetrieveFromCache(true) ).load(it.coverUrl).signature(ObjectKey("covers")).submit()
                 if (future.get().exists())future.get().delete()
                 Glide.with(context).clear(future)
+                }
                 inBookshelf = false
                 if (it.isLocal) {
                     LocalBook.deleteBook(it, deleteOriginal)
